requirements.txt
certifi==2019.11.28
chardet==3.0.4
fonttools==3.44.0
idna==2.8
lxml==4.4.2
PyExecJS==1.5.1
requests==2.22.0
six==1.14.0
urllib3==1.25.7
